requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
click>=8.0
rich>=13.0
pytest>=8.0
pytest-cov>=4.0
